Move alignDataToBoundaries function to utils
