+++ components/DDWDataLoader/DDWDataLoader.tsx
@@ -14,10 +14,9 @@ import {
   GET_INDICATOR_DATA,
   LocationIndicatorData,
   SpotlightBoundary,
-  toCamelCase,
 } from '../../utils';
 import { Alert } from '../Alert';
-import { getLocationIDFromGeoCode } from '../SpotlightMap/utils';
+import { alignDataToBoundaries } from './utils';
 
 export interface DataLoaderProps {
   indicators?: string[];
@@ -28,40 +27,6 @@ export interface DataLoaderProps {
   filter?: DataFilter[][];
   onLoad?: (data: LocationIndicatorData[]) => void;
 }
-
-const alignDataToBoundaries = (
-  data: LocationIndicatorData[],
-  boundaries: SpotlightBoundary[],
-  year?: number
-): LocationIndicatorData[] => {
-  return data.map((indicator) => {
-    const indicatorData = indicator.data.slice();
-    if (indicatorData.length < boundaries.length) {
-      const missingLocations = boundaries.filter((boundary) => {
-        const missing = !indicatorData.find((d) => boundary.geocode.includes(d.geocode));
-
-        return year && missing ? missing && parseInt(boundary.created || '0') > year : missing;
-      });
-
-      missingLocations
-        .filter((d) => d.parent)
-        .forEach((boundary) => {
-          const parent = indicatorData.find((d) => boundary.parent?.includes(d.geocode));
-          if (parent) {
-            const location = {
-              ...parent,
-              geocode: getLocationIDFromGeoCode(boundary.geocode, '.'),
-              name: toCamelCase(boundary.name),
-            };
-
-            indicatorData.push(location);
-          }
-        });
-    }
-
-    return { ...indicator, data: indicatorData };
-  });
-};
 
 const DDWDataLoader: FunctionComponent<DataLoaderProps & { boundaries: SpotlightBoundary[] }> = ({
   indicators,
